Refactor chat and messages relationships & refactor resolvers structure
