+++ api/src/gql/schemes/Chat.gql
@@ -1,8 +1,17 @@
+type Message {
+  id: ID!
+  content: String!
+  author: User!
+  chatId: ID!
+  createdAt: String!
+  lastEdit: String
+}
+
 type Chat {
   id: ID!
   participants: [User]!
   logo: String
-  messages: [Message]
+  messages: [Message] # pagination will live here (it will become a sub-query)
   settings: ChatSettings
 }
 
@@ -17,8 +26,10 @@ type Query {
 
 type Mutation {
   createChat: Chat!
+  createMessage(content: String!, chatId: String!): Message!
 }
 
 type Subscription {
   newChat: Chat!
+  newMessage: Message!
 }
