--- api/src/gql/schemes/Message.gql
@@ -1,20 +1,0 @@
-type Message {
-  id: ID!
-  content: String!
-  author: User!
-  chatId: ID!
-  createdAt: String!
-  lastEdit: String
-}
-
-type Query {
-  getMessages(chatId: ID!): [Message]
-}
-
-type Mutation {
-  createMessage(content: String!, chatId: String!): Message!
-}
-
-type Subscription {
-  newMessage: Message!
-}
