+++ api/src/models/Chat/index.ts
@@ -6,11 +6,6 @@ import {
 } from '@typegoose/typegoose';
 import { ID, Nullable, URL } from '../../types/common';
 
-export type MessageType = {
-  _id: ID;
-  author: ID;
-};
-
 @modelOptions({
   options: { allowMixed: Severity.ALLOW },
   schemaOptions: { collection: 'chats', _id: true },
@@ -24,12 +19,9 @@ export class ChatEntity {
   @prop({ required: true })
   participants: ID[];
 
-  @prop()
-  messages: MessageType[];
-
   constructor(participants: ID[]) {
     this.participants = participants;
-    this.messages = [];
+    this.logo = null;
   }
 }
