--- api/src/resolvers/Message/index.ts
@@ -1,6 +1,0 @@
-import { MessageResolvers } from '../../types/gql.generated';
-
-// TODO: implement friend population
-const messageResolvers: MessageResolvers = {};
-
-export default messageResolvers;
