--- api/src/resolvers/User/index.ts
@@ -1,6 +1,0 @@
-import { UserResolvers } from '../../types/gql.generated';
-
-// TODO: implement friend population
-const userResolvers: UserResolvers = {};
-
-export default userResolvers;
